requirements.txt
airtop==0.0.49
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
boto3==1.34.88
beautifulsoup4==4.13.4
behave==1.2.6
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
execnet==2.1.1
executing==2.2.0
fasteners==0.19
filelock==3.18.0
html2text==2024.2.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mycdp==1.2.0
nest-asyncio==1.6.0
outcome==1.3.0.post0
openai==0.28.0
phonenumbers==8.13.35
pymongo==4.6.3
PyVirtualDisplay==3.0
python-dotenv
packaging==25.0
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
pdbp==1.7.0
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
pyee==12.0.0
Pygments==2.19.1
pynose==1.5.4
pyotp==2.9.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.1
pytest-xdist==3.7.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
sbvirtualdisplay==1.4.0
selenium
seleniumbase
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
tabcompleter==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0